+++ CHANGELOG.md
@@ -24,4 +24,5 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 - fix game version detection for wrath (beta5)
 - add Fishing profession (beta6)
 - add Heirlooms into Collections (beta6)
+- add Emblem and Dalaran Vendors (beta6)
 - #222 add raw gems into jewelcrafting (beta6)